% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{c,tagged-method}
\alias{c,tagged-method}
\title{Concatenate tagged}
\usage{
\S4method{c}{tagged}(x, ..., recursive = FALSE, use.names = TRUE)
}
\arguments{
\item{x}{a tagged object}

\item{...}{more tagged objects}

\item{recursive}{ignored}

\item{use.names}{ignored}
}
\description{
Concatenate tagged
}
\keyword{internal}
