% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{col_positions}
\alias{col_positions}
\title{For each row in a matrix, identify which columns have a}
\usage{
col_positions(x)
}
\arguments{
\item{x}{a matrix of 0s and 1s}
}
\value{
a list of \code{nrow(x)} where each element is an integer vector of
     column positions that have a 1 for each matrix row.
}
\description{
For each row in a matrix, identify which columns have a
}
\keyword{internal}
