% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_index}
\alias{create_index}
\title{Create an index for a grouping based on the variables available for the gofl}
\usage{
create_index(grouping, data)
}
\arguments{
\item{grouping}{a named list of variable values that define a grouping. The
names of this list should also be names in \code{data}}

\item{data}{a named list of all valid variable values that can be used to
define a grouping. The names of this list are the variable names}
}
\description{
Create an index for a grouping based on the variables available for the gofl
}
\keyword{internal}
