% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDSL.R
\docType{package}
\name{filterDSL-package}
\alias{filterDSL}
\alias{filterDSL-package}
\title{Filtration DSL}
\description{
A simple domain specific language for defining subgroups for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.novisci.com/nsStat/filterDSL}
}

}
\author{
\strong{Maintainer}: Bradley Saul \email{bsaul@novisci.com}

}
