% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofl.R
\docType{package}
\name{gofl-package}
\alias{gofl}
\alias{gofl-package}
\title{gofl}
\description{
A simple domain specific language for grouping, organizing, and filtering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.novisci.com/nsStat/gofl}
}

}
\author{
\strong{Maintainer}: Bradley Saul \email{bsaul@novisci.com}

}
