% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reorder_columns_by_names}
\alias{reorder_columns_by_names}
\title{Reorder a matrix with column names by a character vector}
\usage{
reorder_columns_by_names(mat, names)
}
\arguments{
\item{mat}{a \code{matrix} with \code{colnames} that may contain \code{sep}}

\item{names}{a \code{character} vector by which to reorder the columns.
hese are the grouping keys of the columns (the string before \code{sep})}
}
\value{
a \code{matrix} with the same data but reordered columns
}
\description{
Reorder a matrix with column names by a character vector
}
\keyword{internal}
