% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{create_grouping_matrix}
\alias{create_grouping_matrix}
\title{Creating a grouping matrix}
\usage{
create_grouping_matrix(formula, data)
}
\arguments{
\item{formula}{a RHS formula}

\item{data}{list where each element is a}
}
\description{
Creating a grouping matrix
}
