% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofl.R
\name{create_groupings}
\alias{create_groupings}
\title{Create a list of filtration quosures}
\usage{
create_groupings(formula, data, include_overall = FALSE)
}
\arguments{
\item{formula}{a RHS formula}

\item{data}{list where each element corresponds to a term in the \code{formula}}

\item{include_overall}{an indicator of whether to include an "overall" grouping
(i.e. a row in the matrix with no 1s).}
}
\description{
Each element is a list of quosures to be used in a filter statement. The list
of quosures that can be used in a \code{\link[dplyr]{filter}} call, e.g.,
\code{filter(!!! flist)}.
}
